Source files for room, door, maze, and a driver
Some first draft hack and slash source files for how the door, room, and
maze objects will look, as well as a driver which is used for loading a
maze from an input file formatted with the rules laid out in the room
input format for maze text file. There is also a tester I used basically
just to make sure the maze was loading correctly and each room was being
assigned the doors which it should have.