Added some functionality for getting questions
Questions being asked were added to each gui question object, as well as
an int to determine if the question has been answered yet and if it was
answered correctly. added the code in maze driver to create random
question gui and wait for it to be answered, as well as a loop so if it
is answered incorrectly you remain in the same room and it will be
reprinted.